backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
firebase-admin==6.9.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.10.1
pyparsing==3.2.3
python-decouple==3.8
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
